% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RColorLisa.R
\name{color.lisa.pal}
\alias{color.lisa.pal}
\title{Color Palettes from the Masters}
\usage{
color.lisa.pal(n = 5, name)
}
\arguments{
\item{n}{Number of colors to return.}

\item{name}{Name of painting}
}
\description{
This function allows you to load color palettes from great painters.
}
\examples{
color.lisa.pal()
}
\keyword{colors}

